Configure dialyzer and fix multiple warnings
* This conventions the use of the `binary` type rather than `String.t`
* There is still quite a lot of noise cause by -Wunmatched_returns, but I think
we should keep it. These warnings are annoying when we are just interested on
the side effects of a function call. However, it's important to know when we are
not matching on returns as sometimes we should have done it. I have an idea of
how to integrate these reports into PRs and care just about the new warnings.
* For now, I'm ignoring some warnings related to:
  * guards that actually don't exist on the lines dialyzer points to
  * string interpolation that work fine